animation and performance/bug fixes
