+++ src/components/cards/entry-card.tsx
@@ -1,6 +1,7 @@
 import Link from "next/link";
 import toast from "react-hot-toast";
 import { HiX } from "react-icons/hi";
+import { LazyMotion, m, domAnimation } from "framer-motion";
 import { InferQueryOutput, trpc } from "@/utils/trpc";
 import { useProfileStore } from "@/utils/profile";
 
@@ -48,66 +49,72 @@ export default function EntryCard({ entry, page }: Props) {
   };
 
   return (
-    <article
-      key={entry.id}
-      className='px-5 py-4 space-y-2 rounded-md bg-zinc-900 border border-zinc-500'
-    >
-      <div className='flex flex-wrap gap-2 items-center justify-between'>
-        <p className='text-xl font-semibold'>
-          {entry.createdAt.toLocaleDateString()}
-        </p>
+    <LazyMotion features={domAnimation}>
+      <m.article
+        initial={{ opacity: 0, y: 50 }}
+        animate={{ opacity: 1, y: 0 }}
+        transition={{ duration: 0.3 }}
+        className='px-5 py-4 space-y-2 rounded-md bg-zinc-900 border border-zinc-500'
+      >
+        <div className='flex flex-wrap gap-2 items-center justify-between'>
+          <p className='text-xl font-semibold'>
+            {entry.createdAt.toLocaleDateString()}
+          </p>
 
-        <div className='flex flex-wrap gap-2'>
-          <Link
-            className='button-edit'
-            href={`/entry/${entry.id}/edit?page=${
-              page === "workout" ? "workout" : "exercise"
-            }`}
-          >
-            Edit
-          </Link>
-
-          <button
-            className='button-remove'
-            onClick={handleDelete}
-            disabled={isLoading}
-          >
-            <HiX className='h-5 w-5' />
-            <p>Delete</p>
-          </button>
-        </div>
-      </div>
-
-      {entry?.exercise && (
-        <p className='truncate'>
-          <strong>Exercise:</strong> {entry?.exercise.name}
-        </p>
-      )}
-
-      {entry?.notes && (
-        <p className='line-clamp-3'>
-          <strong>Notes:</strong> {entry.notes}
-        </p>
-      )}
-
-      <ul>
-        {entry.sets?.map(set => (
-          <li key={set.id}>
-            {set.weight && (
-              <>
-                {weightUnit === "kg" && `${set.weight} kg - ${set.reps} reps`}
-                {weightUnit === "lbs" &&
-                  `${convertKilosToPounds(set.weight)} lbs - ${set.reps} reps`}
-              </>
-            )}
-
-            {set.distance &&
-              `${set.distance} miles ${
-                set.elevation ? `- ${set.elevation} ft` : ""
+          <div className='flex flex-wrap gap-2'>
+            <Link
+              className='button-edit'
+              href={`/entry/${entry.id}/edit?page=${
+                page === "workout" ? "workout" : "exercise"
               }`}
-          </li>
-        ))}
-      </ul>
-    </article>
+            >
+              Edit
+            </Link>
+
+            <button
+              className='button-remove'
+              onClick={handleDelete}
+              disabled={isLoading}
+            >
+              <HiX className='h-5 w-5' />
+              <p>Delete</p>
+            </button>
+          </div>
+        </div>
+
+        {entry?.exercise && (
+          <p className='truncate'>
+            <strong>Exercise:</strong> {entry?.exercise.name}
+          </p>
+        )}
+
+        {entry?.notes && (
+          <p className='line-clamp-3'>
+            <strong>Notes:</strong> {entry.notes}
+          </p>
+        )}
+
+        <ul>
+          {entry.sets?.map(set => (
+            <li key={set.id}>
+              {set.weight && (
+                <>
+                  {weightUnit === "kg" && `${set.weight} kg - ${set.reps} reps`}
+                  {weightUnit === "lbs" &&
+                    `${convertKilosToPounds(set.weight)} lbs - ${
+                      set.reps
+                    } reps`}
+                </>
+              )}
+
+              {set.distance &&
+                `${set.distance} miles ${
+                  set.elevation ? `- ${set.elevation} ft` : ""
+                }`}
+            </li>
+          ))}
+        </ul>
+      </m.article>
+    </LazyMotion>
   );
 }
