+++ src/pages/workout/[id].tsx
@@ -117,7 +117,7 @@ const Workout = ({ workoutId }: Props) => {
       <DateBar date={date} setDate={setDate} />
 
       <section className='space-y-4'>
-        {filteredData && filteredData?.entries.length <= 0 && (
+        {filteredData && filteredData.length === 0 && (
           <EmptyCard>No entries found with this date</EmptyCard>
         )}
